requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Hugging Face (NEW)
transformers>=4.35.0
datasets>=2.15.0
torch>=2.0.0
accelerate>=0.25.0

# Sentiment Analysis (keep VADER as backup)
vaderSentiment>=3.3.2

# Data Collection (now optional with HF datasets)
yfinance>=0.2.28
newsapi-python>=0.2.7
finnhub-python>=2.4.18

# API & Web
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
joblib>=1.3.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
